requirements.txt
boto3==1.35.72
botocore==1.35.72
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
boto3==1.35.72
botocore==1.35.72
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
